Python modules renamed to follow PEP 8 style guidelines.
