+++ coherence/logging/console_logging.py
